Move vector related APIs into own module
